% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-c14.R
\name{c14_calibrate}
\alias{c14_calibrate}
\title{Calibrate radiocarbon dates}
\usage{
c14_calibrate(cra, error, ..., engine = c("rcarbon", "OxCal", "Bchron"))
}
\arguments{
\item{cra}{A vector of uncalibrated radiocarbon ages.}

\item{error}{A vector of standard errors associated with \code{cra}}

\item{...}{Optional arguments passed to calibration function.}

\item{engine}{Package to use for calibration, i.e. \code{\link[rcarbon:calibrate]{rcarbon::calibrate()}}
(\code{"rcarbon"}), \code{\link[oxcAAR:oxcalCalibrate]{oxcAAR::oxcalCalibrate()}} (\code{"OxCal"}), or
\code{\link[Bchron:BchronCalibrate]{Bchron::BchronCalibrate()}} (\code{"Bchron"}). Default: \code{"rcarbon"}.}
}
\value{
A list of \code{cal} objects.
}
\description{
A thin wrapper of \code{\link[rcarbon:calibrate]{rcarbon::calibrate()}} that returns calibrated dates as a
single list rather than a `CalDates`` object. Useful, for example, if you want
to add a column of calibrated dates to an existing table with \code{\link[dplyr:mutate]{dplyr::mutate()}}
}
\examples{
data("shub1_radiocarbon")
shub1_radiocarbon \%>\%
  dplyr::mutate(cal = c14_calibrate(cra, error))
}
\seealso{
Other tidy radiocarbon functions:
\code{\link[=c14_sum]{c14_sum()}},
\code{\link[=cal]{cal()}}
}
\concept{tidy radiocarbon functions}
