% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql.R
\name{cql_phase}
\alias{cql_phase}
\title{Describe an unordered group in CQL}
\usage{
cql_phase(name, cql)
}
\arguments{
\item{name}{Character. Label for the phase.}

\item{cql}{Vector of \code{cql} objects contained in the phase.}
}
\value{
A \code{cql} object.
}
\description{
The CQL command \code{Phase} is used to describe an unordered group of events in
a model.
}
\examples{
cql(
  cql_phase("Phase A", c(
    cql_r_date("ABC-001", 10100, 50),
    cql_r_date("ABC-002", 10200, 50),
    cql_r_date("ABC-003", 10300, 50)
  ))
)
}
\references{
\url{https://c14.arch.ox.ac.uk/oxcalhelp/hlp_commands.html}
}
\seealso{
Other CQL functions:
\code{\link[=cql]{cql()}},
\code{\link[=cql_boundary]{cql_boundary()}},
\code{\link[=cql_n]{cql_n()}},
\code{\link[=cql_options]{cql_options()}},
\code{\link{cql_other}},
\code{\link[=cql_r_date]{cql_r_date()}},
\code{\link[=cql_sequence]{cql_sequence()}},
\code{\link[=write_oxcal]{write_oxcal()}}
}
\concept{CQL functions}
