% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql.R
\name{cql_sequence}
\alias{cql_sequence}
\alias{cql_d_sequence}
\alias{cql_p_sequence}
\alias{cql_u_sequence}
\alias{cql_v_sequence}
\title{Describe an ordered group in CQL}
\usage{
cql_sequence(name, cql, boundaries = FALSE)

cql_d_sequence()

cql_p_sequence()

cql_u_sequence()

cql_v_sequence()
}
\arguments{
\item{name}{Character. Label for the sequence.}

\item{cql}{Vector of \code{cql} objects contained in the sequence.}

\item{boundaries}{Logical. If \code{TRUE}, adds \code{Boundary} constraints between
each item in the sequence.}
}
\value{
A \code{cql} object.
}
\description{
The CQL command \code{Sequence} is used to describe an \emph{ordered} group of events
in a model. Specific types of sequence, i.e. \code{D_Sequence}, \code{P_Sequence},
\code{U_Sequence}, and \code{V_Sequence}, are not yet implemented in stratigraphr.
}
\details{
OxCal expects events within sequences to be in chronological order (oldest to
youngest) not stratigraphic order (youngest to oldest).
}
\examples{
cql(
  cql_sequence("Sequence A", c(
    cql_r_date("ABC-001", 10100, 50),
    cql_r_date("ABC-002", 10200, 50),
    cql_r_date("ABC-003", 10300, 50)
  ))
)
}
\references{
\url{https://c14.arch.ox.ac.uk/oxcalhelp/hlp_commands.html}
}
\seealso{
Other CQL functions:
\code{\link[=cql]{cql()}},
\code{\link[=cql_boundary]{cql_boundary()}},
\code{\link[=cql_n]{cql_n()}},
\code{\link[=cql_options]{cql_options()}},
\code{\link{cql_other}},
\code{\link[=cql_phase]{cql_phase()}},
\code{\link[=cql_r_date]{cql_r_date()}},
\code{\link[=write_oxcal]{write_oxcal()}}
}
\concept{CQL functions}
