% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{harris}
\alias{harris}
\title{Construct a Harris matrix from a stratigraphic sequence.}
\usage{
harris(data, context, relation)
}
\arguments{
\item{data}{`tibble` or `data.frame` of contexts and the stratigraphic
relationships between them.}

\item{context}{`character`. Name of the column containing context labels (nodes)}

\item{relation}{`character`. Name of the column describing relations (edges)
between contexts}
}
\value{
A `tibble` of edges with `to` and `from` columns (see [tidygraph::tbl_graph()])
}
\description{
Construct a Harris matrix from a stratigraphic sequence.
}
\examples{
data(harris12)
matrix <- harris(harris12, "context", "above")
plot(matrix)
}
