% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{harris12}
\alias{harris12}
\title{Stratigraphy from Harris Figure 12}
\format{
A data frame with 10 rows and 4 variables:
\describe{
\item{context}{Label of the stratigraphic unit.}
\item{above}{Unit stratigraphically above.}
\item{below}{Unit stratigraphically below.}
\item{equal}{Unit stratigraphically equal.}
}
}
\source{
\insertCite{Harris1979-qs;textual}{stratigraphr}, p. 39.
}
\usage{
harris12
}
\description{
A model stratigraphic sequence used by Edward
\insertCite{Harris1979-qs;textual}{stratigraphr} to illustrate the
construction of a Harris matrix.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
