% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{is_valid_harris}
\alias{is_valid_harris}
\title{Validate stratigraphic graphs}
\usage{
is_valid_harris(stratigraph, warn = TRUE)
}
\arguments{
\item{stratigraph}{A \code{stratigraph} object (see \code{\link[=stratigraph]{stratigraph()}})}

\item{warn}{Display warnings for invalid graphs (Default: \code{TRUE}).}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Checks whether a stratigraphic graph is a valid Harris matrix.
}
