% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{lst_extract_attribute}
\alias{lst_extract_attribute}
\title{Extract a single attribute from an LST file}
\usage{
lst_extract_attribute(attr)
}
\arguments{
\item{attr}{A single attribute line from an LST file.}
}
\value{
A named list containing the attribute name and value.
}
\description{
Extract a single attribute from an LST file
}
