% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{lst_extract_header}
\alias{lst_extract_header}
\title{Extract header from an LST file}
\usage{
lst_extract_header(lst)
}
\arguments{
\item{lst}{Character vector of lines from an LST file (e.g. from \code{\link[readr:read_lines]{readr::read_lines()}})}
}
\description{
Extract header from an LST file
}
