% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{lst_extract_stratum}
\alias{lst_extract_stratum}
\title{Extract a single stratum from an LST file}
\usage{
lst_extract_stratum(stratum)
}
\arguments{
\item{lst}{Character vector of lines from an LST file (e.g. from \code{\link[readr:read_lines]{readr::read_lines()}})
representing a single stratum.}
}
\value{
Data frame with one row and columns representing the name and attributes of
a stratum.
}
\description{
Extract a single stratum from an LST file
}
\details{
Currently lst_extract_stratum() assumes each stratum is exactly five lines
long: the first line with the name of the stratum, and subsequent lines
containing its attributes (passed to \code{\link[=lst_extract_attribute]{lst_extract_attribute()}}). This
may not be a sound assumption for the "extended" format mentioned in
\url{http://archaeologic.al/wiki/Harris_Matrix#LST}.
}
