% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_model.R
\name{model_phase_cql}
\alias{model_phase_cql}
\title{Generate a CQL phase model}
\usage{
model_phase_cql(
  data,
  phase,
  label,
  c14_age,
  c14_error,
  sequence = c("contiguous", "sequential", "overlapping"),
  transition = c("gradual", "abrupt")
)
}
\arguments{
\item{data}{A data frame of radiocarbon-dated events.}

\item{phase}{<\link[dplyr:dplyr_tidy_select]{tidy_select}> Variable(s) used to
group events into phases. Ideally these should be factors with levels
ordered chronologically (oldest to youngest). Otherwise, it is assumed that
they have a natural order, e.g. \verb{1..3} or \code{A..C}.}

\item{label}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Variable to use as
labels for the events, e.g. lab numbers.}

\item{c14_age}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Variable
describing the uncalibrated radiocarbon ages of events.}

\item{c14_error}{<\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Variable
describing the radiocarbon error associated with the events.}

\item{sequence}{Character. Prior assumption of the ordering of phases. One of:
\itemize{
\item \code{"contiguous"} – phases are assumed to have occurred immediately one after
the other
\item \code{"sequential"} – phases are assumed to have occurred one after the other,
with the possibility of a gap between
\item \code{"overlapping"} – phases are assumed to be independent of each other
}}

\item{transition}{Character. Prior assumption of the boundaries of phases.
One of:
\itemize{
\item \code{"gradual"} – assumes a smooth transition at the start and end of each
phase, using a 'trapezium' prior \insertCite{Lee2012}{stratigraphr}.
\item \code{"abrupt"} – assumes an immediate transition at the start and end of each
phase
}}
}
\value{
A CQL script (see \code{\link[=cql]{cql()}}).
}
\description{
Constructs a chronological model from a table of radiocarbon dates and
phases.

Phase models group \emph{events} (in this case radiocarbon dates) into one or more
\emph{phases} that form an ordered sequence, but where the order within a phase is
unknown. The transition between phases is described by a combination of
\emph{boundary} constraints which determine the prior distribution for Bayesian
calibration.
}
\details{
This function is based on the phase modelling approach used in OxCal
\insertCite{Bronk_Ramsey2013}{stratigraphr}. Currently only uniform prior
distributions (using \code{\link[=cql_boundary]{cql_boundary()}}) for phases are supported.
}
\references{
\insertAllCited{}
}
\concept{stratigraphic model functions}
