% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat.R
\name{strat_connect}
\alias{strat_connect}
\title{Connect stratigraphic units}
\usage{
strat_connect(units, relations, direction = c("above", "below"))
}
\arguments{
\item{units}{Vector of unit labels.}

\item{relations}{Vector of relations.}

\item{direction}{Are the units in \code{relations} "above" or "below" the ones
in \code{units}?}
}
\value{
A data frame of directed edges represented by \code{to} and \code{from} columns, which
can be used as the \code{nodes} argument to \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph()}}.
}
\description{
Constructs a table of directed edges between stratigraphic units based on a
vector of relations (e.g. "above" or "below").
}
\examples{
strat_connect(
  units = c("A", "B", "C"),
  relations = list("B", "C", NA),
  direction = "above"
)
}
