% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcarbon.R
\name{sum_radiocarbon}
\alias{sum_radiocarbon}
\title{Sum radiocarbon dates with tidy syntax}
\usage{
sum_radiocarbon(cal_dates, time_range = NULL, ...)
}
\arguments{
\item{cal_dates}{A list of \code{calGrid} objects.}

\item{time_range}{Vector of length 2 indicating the range of calendar dates
over which to sum. If \code{NULL}, the maximum range of the
\code{cal_dates} will be used.}

\item{...}{Optional arguments to be passed to \code{\link[rcarbon:spd]{rcarbon::spd()}}}
}
\value{
A \code{calGrid} object containing the summed probability distribution.
}
\description{
A wrapper for \code{\link[rcarbon:spd]{rcarbon::spd()}} that takes a list of calibrated dates rather
than a \code{CalDates} object. This allows you to use the output of \code{\link[=cal]{cal()}} and
take advantage of tidy summary syntax (\url{dplyr::group_by}, etc.)
}
\note{
Unlike \code{\link[rcarbon:spd]{rcarbon::spd()}}, this function will attempt to guess the time range
if it isn't specified. It's probably a good idea to specify it.
}
\examples{
library("tidyverse")
library("rcarbon")
data("emedyd")
emedyd \%>\%
  mutate(CalDate = cal(CRA, Error, normalise = FALSE)) ->
  emedyd
emedyd \%>\%
  group_by(SiteName) \%>\%
  summarise(SPD = sum_radiocarbon(CalDate, spdnormalised = TRUE)) ->
  emedyd_summed
}
