% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-cal.R
\name{as.CalDates.cal}
\alias{as.CalDates.cal}
\title{Convert cal objects to a rcarbon CalDates object}
\usage{
as.CalDates.cal(x)
}
\arguments{
\item{x}{A list of \code{cal} objects.}
}
\value{
A \code{CalDates} object. See \code{\link[rcarbon:calibrate]{rcarbon::calibrate()}} for details.
}
\description{
Convert cal objects to a rcarbon CalDates object
}
\seealso{
Other functions for working with \code{cal} objects: 
\code{\link{as_cal}()},
\code{\link{cal_metadata}()},
\code{\link{cal}()}
}
\concept{functions for working with \code{cal} objects}
