% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\name{as_cal}
\alias{as_cal}
\alias{as_cal.CalDates}
\alias{as_cal.oxcAARCalibratedDatesList}
\alias{as_cal.oxcAARCalibratedDate}
\alias{as_cal.BchronCalibratedDates}
\title{Convert an object to a cal object}
\usage{
as_cal(x)

\method{as_cal}{CalDates}(x)

\method{as_cal}{oxcAARCalibratedDatesList}(x)

\method{as_cal}{oxcAARCalibratedDate}(x)

\method{as_cal}{BchronCalibratedDates}(x)
}
\arguments{
\item{x}{object to be converted to a \code{cal} object.}
}
\value{
\code{cal} object: a data frame with two columns, \code{year} and \code{p}, representing
the calibrated probability distribution. All other values are stored as
attributes and can be accessed with \code{\link[=cal_metadata]{cal_metadata()}}.
}
\description{
\code{as_cal()} converts objects from other packages that represent calibrated
radiocarbon dates to \code{cal} objects.
Methods are currently implemented for:
\itemize{
\item \code{CalDates}: from \code{\link[rcarbon:calibrate]{rcarbon::calibrate()}}
\item \code{oxcAARCalibratedDate} and \code{oxcAARCalibratedDatesList}: from \code{\link[oxcAAR:oxcalCalibrate]{oxcAAR::oxcalCalibrate()}}
\item \code{BchronCalibratedDates}: from \code{\link[Bchron:BchronCalibrate]{Bchron::BchronCalibrate()}}
}

These functions are intended for complex S3 objects from other packages.
See \code{\link[=cal]{cal()}} for a more generic constructor, e.g. using a data frame.
}
\seealso{
Other functions for working with \code{cal} objects: 
\code{\link{as.CalDates.cal}()},
\code{\link{cal_metadata}()},
\code{\link{cal}()}
}
\concept{functions for working with \code{cal} objects}
