% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\name{as_cal}
\alias{as_cal}
\alias{as_cal.CalDates}
\title{Convert an object to a cal object}
\usage{
as_cal(x)

\method{as_cal}{CalDates}(x)
}
\arguments{
\item{x}{object to be converted to a \code{cal} object.}
}
\value{
\code{cal} object: a data frame with two columns, \code{year} and \code{p}, representing
the calibrated probability distribution. All other values are stored as
attributes.
}
\description{
\code{as_cal()} converts objects from other packages that represent calibrated
radiocarbon dates to \code{cal} objects.
Methods are currently implemented for:
\itemize{
\item \code{CalDates}: from \code{\link[rcarbon:calibrate]{rcarbon::calibrate()}}
}

These functions are intended for complex S3 objects from other packages.
See \code{\link[=cal]{cal()}} for a more generic constructor, e.g. using a data frame.
}
\seealso{
Other radiocarbon functions: 
\code{\link{c14_calibrate}()},
\code{\link{c14_sum}()},
\code{\link{cal}()}
}
\concept{radiocarbon functions}
