% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcarbon.R
\name{c14_calibrate}
\alias{c14_calibrate}
\title{Calibrate radiocarbon dates with tidy syntax}
\usage{
c14_calibrate(cra, error, ...)
}
\arguments{
\item{cra}{A vector of uncalibrated radiocarbon ages.}

\item{error}{A vector of standard errors associated with \code{cra}}

\item{...}{Optional arguments passed to \code{\link[rcarbon:calibrate]{rcarbon::calibrate()}}}
}
\value{
A list of \code{calGrid} objects. See \code{\link[rcarbon:calibrate]{rcarbon::calibrate()}} for details.
}
\description{
A thin wrapper of \code{\link[rcarbon:calibrate]{rcarbon::calibrate()}} that returns calibrated dates as a
single list rather than a `CalDates`` object. Useful, for example, if you want
to add a column of calibrated dates to an existing table with \code{\link[dplyr:mutate]{dplyr::mutate()}}
}
\note{
This function only supports the 'grid' method of representing calibrated
dates (i.e. \code{calMatrix} is always set to \code{FALSE}) and the metadata usually
returned by \code{\link[rcarbon:calibrate]{rcarbon::calibrate()}} is discarded.
}
\examples{
data("shub1_radiocarbon")
shub1_radiocarbon \%>\%
  dplyr::mutate(CalDate = c14_calibrate(cra, error, normalise = FALSE, verbose = FALSE))
}
\seealso{
Other radiocarbon functions: 
\code{\link{as_cal}()},
\code{\link{c14_sum}()},
\code{\link{cal}()}
}
\concept{radiocarbon functions}
