% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\name{cal}
\alias{cal}
\alias{print.cal}
\title{Calibrated radiocarbon dates}
\usage{
cal(
  x,
  era = c("cal BP"),
  lab_id = NA,
  cra = NA,
  error = NA,
  curve = NA,
  reservoir_offset = NA,
  reservoir_offset_error = NA,
  calibration_range = NA,
  F14C = NA,
  normalised = NA,
  p_cutoff = NA
)

\method{print}{cal}(x, ...)
}
\arguments{
\item{x}{A \code{data.frame} with two columns, with calendar years and associated probabilities.}

\item{era}{(Optional) \code{character}. Calendar system used. Default: \verb{cal BP}.}

\item{lab_id}{(Optional) \code{character}. Lab code or other label for the calibrated sample.}

\item{cra}{(Optional) \code{integer}. Uncalibrated conventional radiocarbon age (CRA) of the sample.}

\item{error}{(Optional) \code{integer}. Error associated with the uncalibrated sample.}

\item{curve}{(Optional) \code{character}. Atmospheric curve used for calibration, e.g. "intcal20".}

\item{reservoir_offset}{(Optional) \code{integer}. Marine reservoir offset used in the calibration, if any.}

\item{reservoir_offset_error}{(Optional) \code{integer}. Error associated with the marine reservoir offset.}

\item{calibration_range}{(Optional) \code{integer} vector of length 2. The range of years over which the calibration was performed, i.e. \code{c(start, end)}.}

\item{F14C}{(Optional) \code{logical}. Whether the calibration was calculated using F14C values instead of the CRA.}

\item{normalised}{(Optional) \code{logical}. Whether the calibrated probability densities were normalised.}

\item{p_cutoff}{(Optional) \code{numeric}. Lower threshold beyond which probability densities were considered zero.}

\item{...}{(Optional) Arguments based to other functions.}
}
\value{
\code{cal} object: a data frame with two columns, \code{year} and \code{p}, representing
the calibrated probability distribution. All other values are stored as
attributes and can be accessed with \code{\link[=cal_metadata]{cal_metadata()}}.
}
\description{
\code{cal()} constructs a generic representation of a calibrated radiocarbon date
or other calendar-based probability distribution.
}
\seealso{
Other tidy radiocarbon functions: 
\code{\link{c14_calibrate}()},
\code{\link{c14_sum}()}

Other functions for working with \code{cal} objects: 
\code{\link{as.CalDates.cal}()},
\code{\link{as_cal}()},
\code{\link{cal_metadata}()}
}
\concept{functions for working with \code{cal} objects}
\concept{tidy radiocarbon functions}
