% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-cal.R
\name{cal_metadata}
\alias{cal_metadata}
\title{Extract metadata from a calibrated date}
\usage{
cal_metadata(x)
}
\arguments{
\item{x}{A \code{cal} object. See \code{\link[=cal]{cal()}}.}
}
\value{
A named list of metadata attributes.
}
\description{
Extract metadata from a calibrated date
}
\seealso{
Other functions for working with \code{cal} objects: 
\code{\link{as.CalDates.cal}()},
\code{\link{as_cal}()},
\code{\link{cal}()}
}
\concept{functions for working with \code{cal} objects}
