% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql.R
\name{cql}
\alias{cql}
\title{Chronological Query Language (CQL)}
\usage{
cql(...)
}
\arguments{
\item{...}{\code{cql} objects.}
}
\value{
A CQL script.
}
\description{
Provides an R API for the Chronological Query Language (CQL), primarily used
to input commands and describe models to OxCal.
}
\details{
See \code{vignette("cql")} for further details.

List arguments to \code{cql()} are collapsed to produce a single script.
}
\references{
\url{https://c14.arch.ox.ac.uk/oxcalhelp/hlp_commands.html}
}
\seealso{
Other CQL functions: 
\code{\link{cql_boundary}()},
\code{\link{cql_options}()},
\code{\link{cql_other}},
\code{\link{cql_phase}()},
\code{\link{cql_r_date}()},
\code{\link{cql_sequence}()},
\code{\link{write_oxcal}()}
}
\concept{CQL functions}
