% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql.R
\name{cql}
\alias{cql}
\alias{cql_after}
\alias{cql_age}
\alias{cql_axis}
\alias{cql_before}
\alias{cql_boundary}
\alias{cql_c_date}
\alias{cql_c_combine}
\alias{cql_c_simulate}
\alias{cql_combine}
\alias{cql_correl_matrix}
\alias{cql_correlation}
\alias{cql_covar_matrix}
\alias{cql_curve}
\alias{cql_date}
\alias{cql_delta_r}
\alias{cql_d_sequence}
\alias{cql_difference}
\alias{cql_end}
\alias{cql_exp}
\alias{cql_first}
\alias{cql_gap}
\alias{cql_interval}
\alias{cql_kde_model}
\alias{cql_kde_plot}
\alias{cql_label}
\alias{cql_last}
\alias{cql_line}
\alias{cql_lnn}
\alias{cql_mcmc_sample}
\alias{cql_mix_curves}
\alias{cql_n}
\alias{cql_number}
\alias{cql_order}
\alias{cql_offset}
\alias{cql_outlier}
\alias{cql_outlier_model}
\alias{cql_p}
\alias{cql_p_sequence}
\alias{cql_phase}
\alias{cql_pois}
\alias{cql_prior}
\alias{cql_probability}
\alias{cql_r_date}
\alias{cql_r_combine}
\alias{cql_r_f14c}
\alias{cql_r_simulate}
\alias{cql_reservoir}
\alias{cql_sample}
\alias{cql_sapwood}
\alias{cql_sapwood_model}
\alias{cql_sequence}
\alias{cql_sigma_boundary}
\alias{cql_shift}
\alias{cql_span}
\alias{cql_start}
\alias{cql_sum}
\alias{cql_t}
\alias{cql_tau_boundary}
\alias{cql_top_hat}
\alias{cql_transition}
\alias{cql_u}
\alias{cql_u_sequence}
\alias{cql_v_sequence}
\alias{cql_zero_boundary}
\title{Chronological Query Language (CQL)}
\usage{
cql(...)

cql_after()

cql_age()

cql_axis()

cql_before()

cql_boundary(name, likelihood = NULL)

cql_c_date()

cql_c_combine()

cql_c_simulate()

cql_combine()

cql_correl_matrix()

cql_correlation()

cql_covar_matrix()

cql_curve()

cql_date()

cql_delta_r()

cql_d_sequence()

cql_difference()

cql_end()

cql_exp()

cql_first()

cql_gap()

cql_interval()

cql_kde_model()

cql_kde_plot()

cql_label()

cql_last()

cql_line()

cql_lnn()

cql_mcmc_sample()

cql_mix_curves()

cql_n()

cql_number()

cql_order()

cql_offset()

cql_outlier()

cql_outlier_model()

cql_p()

cql_p_sequence()

cql_phase(name, ...)

cql_pois()

cql_prior()

cql_probability()

cql_r_date(name, cra, error)

cql_r_combine()

cql_r_f14c()

cql_r_simulate()

cql_reservoir()

cql_sample()

cql_sapwood()

cql_sapwood_model()

cql_sequence(name, ..., add_boundaries = FALSE)

cql_sigma_boundary()

cql_shift()

cql_span()

cql_start()

cql_sum()

cql_t()

cql_tau_boundary()

cql_top_hat()

cql_transition()

cql_u()

cql_u_sequence()

cql_v_sequence()

cql_zero_boundary()
}
\value{
\code{cql}. A CQL script.
}
\description{
Provides an R API for the Chronological Query Language (CQL), primarily used
to input commands and describe models to OxCal.
}
