% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql.R
\name{cql_boundary}
\alias{cql_boundary}
\alias{cql_sigma_boundary}
\alias{cql_tau_boundary}
\alias{cql_zero_boundary}
\title{Describe a boundary constraint in CQL}
\usage{
cql_boundary(name, likelihood = NULL)

cql_sigma_boundary()

cql_tau_boundary()

cql_zero_boundary()
}
\arguments{
\item{name}{Character. Label for the boundary.}

\item{likelihood}{Expression defining the likelihood of the boundary.}
}
\value{
A \code{cql} object.
}
\description{
The CQL command \code{Boundary} describes constraints within an ordered sequence
(see \code{\link[=cql_sequence]{cql_sequence()}}). Groups of events between two boundaries are assumed
to be sampled from the same prior distribution. \code{Boundary} alone models a
uniform prior. Other types of boundary, i.e. \code{Sigma_Boundary}, \code{Tau_Boundary},
and \code{Zero_Boundary} are not yet implemented in stratigraphr.
}
\examples{

}
\references{
\url{https://c14.arch.ox.ac.uk/oxcalhelp/hlp_commands.html}
}
\seealso{
Other CQL functions: 
\code{\link{cql_options}()},
\code{\link{cql_other}},
\code{\link{cql_phase}()},
\code{\link{cql_r_date}()},
\code{\link{cql_sequence}()},
\code{\link{cql}()},
\code{\link{write_oxcal}()}
}
\concept{CQL functions}
