% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql.R
\name{cql_n}
\alias{cql_n}
\alias{cql_lnn}
\alias{cql_t}
\alias{cql_top_hat}
\alias{cql_u}
\title{Describe distributions in CQL}
\usage{
cql_n(name, mu, sigma, resolution = NULL)

cql_lnn(name, mu, sigma, resolution = NULL)

cql_t(name, freedom, scale = 1, resolution = NULL)

cql_top_hat(name, mid, half_width)

cql_u(name, from, to, resolution = NULL)
}
\arguments{
\item{name}{Character. Label for the distribution.}

\item{mu}{Integer. Mean of a normal or log–normal distribution.}

\item{sigma}{Integer. Standard deviation of a normal or log–normal distribution.}

\item{resolution}{Integer. Resolution of the PDF. Leave \code{NULL} (the default) to use the OxCal default.}

\item{freedom}{Integer. Degrees of freedom of a Student's t distribution.}

\item{scale}{Numeric. Optional scaling parameter for a Student's t distribution.}

\item{mid}{Integer. Centre point of a uniform distribution.}

\item{half_width}{Integer. Half-width of a uniform distribution.}

\item{from}{Integer. Start point of a uniform distribution.}

\item{to}{Integer. End point of a uniform distribution.}
}
\value{
A \code{cql} object.
}
\description{
The CQL commands \code{N}, \code{LnN}, \code{T}, \code{U} and \code{Top_Hat} describe various types
of probability distribution functions. \code{cql_n()} defines a normal distribution
by its mean and standard deviation; \code{cql_lnn()} a log–normal distribution by
its mean and standard deviation; \code{cql_t()} a Student's t distribution by
degrees of freedom; \code{cql_u()} a uniform distribution by its start and end;
\code{cql_top_hat()} a uniform distribution by its central point and width.
}
\examples{
# Describe a date as a uniform distribution
cql_date("U-Date", cql_u("U", -5000, -4000))
}
\references{
\url{https://c14.arch.ox.ac.uk/oxcalhelp/hlp_analysis_inform.html#param}

\url{https://c14.arch.ox.ac.uk/oxcalhelp/hlp_commands.html}
}
\seealso{
Other CQL functions: 
\code{\link{cql_boundary}()},
\code{\link{cql_options}()},
\code{\link{cql_other}},
\code{\link{cql_phase}()},
\code{\link{cql_r_date}()},
\code{\link{cql_sequence}()},
\code{\link{cql}()},
\code{\link{write_oxcal}()}
}
\concept{CQL functions}
