% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql.R
\name{cql_other}
\alias{cql_other}
\alias{cql_age}
\alias{cql_axis}
\alias{cql_c_combine}
\alias{cql_c_simulate}
\alias{cql_correl_matrix}
\alias{cql_correlation}
\alias{cql_covar_matrix}
\alias{cql_curve}
\alias{cql_delta_r}
\alias{cql_difference}
\alias{cql_end}
\alias{cql_exp}
\alias{cql_gap}
\alias{cql_interval}
\alias{cql_kde_model}
\alias{cql_kde_plot}
\alias{cql_label}
\alias{cql_line}
\alias{cql_mcmc_sample}
\alias{cql_mix_curves}
\alias{cql_number}
\alias{cql_offset}
\alias{cql_outlier}
\alias{cql_outlier_model}
\alias{cql_p}
\alias{cql_pois}
\alias{cql_prior}
\alias{cql_probability}
\alias{cql_r_combine}
\alias{cql_r_simulate}
\alias{cql_reservoir}
\alias{cql_sample}
\alias{cql_sapwood}
\alias{cql_sapwood_model}
\alias{cql_shift}
\alias{cql_start}
\alias{cql_transition}
\alias{cql_after}
\alias{cql_before}
\alias{cql_combine}
\alias{cql_first}
\alias{cql_last}
\alias{cql_order}
\alias{cql_span}
\alias{cql_sum}
\title{Other CQL functions (unimplemented)}
\usage{
cql_age(...)

cql_axis(...)

cql_c_combine(...)

cql_c_simulate(...)

cql_correl_matrix(...)

cql_correlation(...)

cql_covar_matrix(...)

cql_curve(...)

cql_delta_r(...)

cql_difference(...)

cql_end(...)

cql_exp(...)

cql_gap(...)

cql_interval(...)

cql_kde_model(...)

cql_kde_plot(...)

cql_label(...)

cql_line(...)

cql_mcmc_sample(...)

cql_mix_curves(...)

cql_number(...)

cql_offset(...)

cql_outlier(...)

cql_outlier_model(...)

cql_p(...)

cql_pois(...)

cql_prior(...)

cql_probability(...)

cql_r_combine(...)

cql_r_simulate(...)

cql_reservoir(...)

cql_sample(...)

cql_sapwood(...)

cql_sapwood_model(...)

cql_shift(...)

cql_start(...)

cql_transition(...)

cql_after()

cql_before()

cql_combine()

cql_first()

cql_last()

cql_order()

cql_span()

cql_sum()
}
\arguments{
\item{...}{Ignored.}
}
\value{
A warning not to use these functions!
}
\description{
These CQL commands are not yet implemented in stratigraphr and only included
here for reference. Unimplemented commands can be included manually using
\code{\link[=cql]{cql()}} (see examples), but if you need one for an analysis please file a bug
report!
}
\examples{
# Use cql() to manually write commands that are not yet implemented, e.g.:
cql("After(\"A\");")
}
\seealso{
Other CQL functions: 
\code{\link{cql_boundary}()},
\code{\link{cql_n}()},
\code{\link{cql_options}()},
\code{\link{cql_phase}()},
\code{\link{cql_r_date}()},
\code{\link{cql_sequence}()},
\code{\link{cql}()},
\code{\link{write_oxcal}()}
}
\concept{CQL functions}
