% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql.R
\name{cql_r_date}
\alias{cql_r_date}
\alias{cql_c_date}
\alias{cql_r_f14c}
\alias{cql_date}
\title{Describe dates in CQL}
\usage{
cql_r_date(name, date, error)

cql_c_date(name, date, error)

cql_r_f14c(name, date, error)

cql_date(name, cql)
}
\arguments{
\item{name}{Character. Date label(s), usually a lab code.}

\item{date}{Integer. Date or dates expressed in radiocarbon years (\code{cql_r_date}),
F14C (\code{cql_f14c_date}) or calendar years (\code{cql_c_date}). See details.}

\item{error}{Integer. Uncertainty associated with the date(s).}

\item{cql}{\code{cql} object. Expression to be converted to a Date.}
}
\value{
A \code{cql} object, or a list of \code{cql} objects if the arguments are vectors.
}
\description{
The CQL commands \code{R_Date}, \code{F14C_Date}, \code{C_Date} and \code{Date} are used to
describe individual dated events in a model. \code{R_Date} and \code{F14C_Date}
represent radiocarbon dates expressed in radiocarbon years (the conventional
radiocarbon age, CRA) or \emph{fraction modern} (F14C) respectively.
\code{C_Date} represents a calendar date. \code{Date} is a type conversion function
which coerces other CQL expressions (e.g. describing a probability
distribution) to a date.
}
\details{
The era expected for calendar dates (BP or BC/AD) depends on a global option
in OxCal, which defaults to BC/AD. BC dates are specified as negative values.

F14C measurements are recommended for modern, "post-bomb" radiocarbon dates
\insertCite{Reimer2004-yl}{stratigraphr}.
}
\references{
\url{https://c14.arch.ox.ac.uk/oxcalhelp/hlp_analysis_inform.html#date}

\url{https://c14.arch.ox.ac.uk/oxcalhelp/hlp_commands.html}

\insertAllCited{}
}
\seealso{
Other CQL functions: 
\code{\link{cql_boundary}()},
\code{\link{cql_n}()},
\code{\link{cql_options}()},
\code{\link{cql_other}},
\code{\link{cql_phase}()},
\code{\link{cql_sequence}()},
\code{\link{cql}()},
\code{\link{write_oxcal}()}
}
\concept{CQL functions}
