% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{harris}
\alias{harris}
\title{Harris Matrices}
\usage{
harris(data, context, relation)
}
\arguments{
\item{data}{\code{tibble} or \code{data.frame} of contexts and the stratigraphic
relationships between them.}

\item{context}{\code{character}. Name of the column containing context labels (nodes)}

\item{relation}{\code{character}. Name of the column describing relations (edges)
between contexts}
}
\value{
A \code{tibble} of edges with \code{to} and \code{from} columns (see \code{\link[tidygraph:tbl_graph]{tidygraph::tbl_graph()}})
}
\description{
Join stratigraphic units using the Harris matrix method (Harris 1979).
}
\examples{
data(harris12)
harris(harris12, "context", "above")
}
\references{
\itemize{
\item Harris, E. C. 1979. \emph{Principles of archaeological stratigraphy}. London:
Academic Press.
}
}
