% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{is_valid_harris}
\alias{is_valid_harris}
\title{Validate stratigraphc graphs}
\usage{
is_valid_harris(stratgraph, warn = TRUE)
}
\arguments{
\item{stratgraph}{A stratigraphic graph.}

\item{warn}{Display warnings for invalid graphs (Default: `TRUE`).}
}
\value{
`TRUE` or `FALSE`
}
\description{
Checks whether a stratigraphic graph is a valid Harris matrix.
}
