% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_lst}
\alias{read_lst}
\title{Read an LST file}
\usage{
read_lst(file, locale = readr::default_locale())
}
\arguments{
\item{file}{Path to an LST file.}

\item{locale}{\code{\link[readr:locale]{readr::locale()}} object specifying the character encoding and
other region-specific settings used in the file. Defaults to
\code{\link[readr:locale]{readr::default_locale()}}, which should be fine in most cases.}
}
\value{
A data frame. Each row represents one stratum. Columns contain the name of
the stratum and the attributes associated with it (typically "above",
"contemporary_with", "equal_to" and "below").
}
\description{
Reads stratigraphic data in LST format (used by BASP Harris, Stratify, and
ArchEd) into a data frame.
}
\references{
\url{http://archaeologic.al/wiki/Harris_Matrix#LST}
}
