% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shub1}
\alias{shub1}
\title{Schematic stratigraphy of Shubayqa 1}
\format{
A data frame with 30 rows, representing contexts, and 7 variables:
\describe{
\item{context}{integer; a unique identifier of the context.}
\item{type}{character; type of context, i.e. deposit, fill, cut,
or structural.}
\item{above}{integer vector; context(s) stratigraphically above this one.}
\item{below}{integer vector; context(s) stratigraphically below this one.}
\item{equal}{integer vector; context(s) stratigraphically
equal to this one.}
\item{phase}{character; Phase assigned to the context,
for contexts that aren't structural.}
\item{structure}{character; for structural contexts,
the name of the structure they belong to.}
}
}
\source{
\insertCite{Richter2017-xy;textual}{stratigraphr}
}
\usage{
shub1
}
\description{
A simplified version of the stratigraphy of Shubayqa 1,
an Epipalaeolithic site in eastern Jordan,
after \insertCite{Richter2017-xy;textual}{stratigraphr}.
}
\details{
The stratigraphy is a simplified version derived from the schematic section
in \insertCite{Richter2017-xy;textual}{stratigraphr}, figure 2.
Context numbers were arbitrarily assigned and the stratigraphic relations are
based on those evident in the diagram.
Phase and structure names are also based on the diagram.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
