% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratigraph.R
\name{strat_is_mirror}
\alias{strat_is_mirror}
\title{Are two relation vectors mirrored?}
\usage{
strat_is_mirror(units, relation1, relation2)
}
\arguments{
\item{units}{Vector of unit labels.}

\item{relation1}{First vector of relations.}

\item{relation2}{Second vector of relations.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Checks whether one vector of relations is the inverse of another. Typically
used to confirm that "above" and "below" columns match and will result in
the same stratigraphic graph.
}
\examples{
data("harris12")
strat_is_mirror(harris12$context, harris12$above, harris12$below)
}
