% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratigraph.R
\name{stratigraph}
\alias{stratigraph}
\title{Construct a stratigraphic graph}
\usage{
stratigraph(data, label, relation, direction = c("above", "below"))
}
\arguments{
\item{data}{Data frame of stratigraphic units, containing at least a unique
label column and a column describing stratigraphic relations.}

\item{label}{Name of the column containing labels of the stratigraphic units.}

\item{relation}{Name of the column describing the stratigraphic relations
between units.}

\item{direction}{Direction described by \code{relation}, i.e. are the units in
that column "above" or "below" the ones in units. Default: "above".}
}
\value{
A \code{stratigraph} object.
}
\description{
Takes a data frame describing stratigraphic units and relations and turns
it into a graph representation.
}
\examples{
circle <- stratigraph(data.frame(
  label = LETTERS[1:4],
  below = c("B", "C", "D", "A")
), "label", "below", "below")
}
