% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{stratigraph}
\alias{stratigraph}
\title{Stratigraphic graphs}
\usage{
stratigraph(data, context, relation, type = "harris")
}
\arguments{
\item{data}{\code{tibble} or \code{data.frame} of contexts and the stratigraphic
relationships between them.}

\item{context}{\code{character}. Name of the column containing context labels (nodes)}

\item{relation}{\code{character}. Name of the column describing relations (edges)
between contexts}

\item{type}{Type of graph. Currently only "harris" is supported.}
}
\value{
A \code{stratigraph} object.
}
\description{
Stratigraphic graphs
}
