% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratigraphr-package.R
\docType{package}
\name{stratigraphr-package}
\alias{stratigraphr}
\alias{stratigraphr-package}
\title{stratigraphr: Archaeological Stratigraphy and Chronological Sequences}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

A tidy framework for working with archaeological stratigraphy and chronology. Includes tools for reading, analysing, and visualising stratigraphic sequences (Harris matrices) as directed graphs following the definition of Dye and Buck (2015) \doi{10.1016/j.jas.2015.08.008} and an R interface to the Chronological Query Language (CQL) used in 'OxCal' by Bronk Ramsey (2009) \doi{10.1017/S0033822200033865}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stratigraphr.joeroe.io}
  \item \url{https://github.com/joeroe/stratigraphr}
  \item Report bugs at \url{https://github.com/joeroe/stratigraphr/issues}
}

}
\author{
\strong{Maintainer}: Joe Roe \email{joe@joeroe.io} (\href{https://orcid.org/0000-0002-1011-1244}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Joe Roe \email{joe@joeroe.io} (\href{https://orcid.org/0000-0002-1011-1244}{ORCID}) [copyright holder]
}

}
\keyword{internal}
