% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratigraphr-package.R
\docType{package}
\name{stratigraphr-package}
\alias{stratigraphr}
\alias{stratigraphr-package}
\title{stratigraphr: Archaeological Stratigraphy and Chronological Sequences}
\description{
A tidy framework for working with archaeological stratigraphy and
    chronology in R. Includes tools for reading, analysing, and visualising
    stratigraphies ("Harris matrices") and sequences as directed graphs; 
    helper functions for using radiocarbon dates in a tidy data analysis; and an
    R interface to OxCal's Chronological Query Language (CQL).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stratigraphr.joeroe.io}
  \item \url{https://github.com/joeroe/stratigraphr}
  \item Report bugs at \url{https://github.com/joeroe/stratigraphr/issues}
}

}
\author{
\strong{Maintainer}: Joe Roe \email{joe@joeroe.io} (\href{https://orcid.org/0000-0002-1011-1244}{ORCID})

}
\keyword{internal}
