% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratigraph.R
\name{strg_is_valid}
\alias{strg_is_valid}
\title{Is an object a valid stratigraphic graph?}
\usage{
strg_is_valid(stratigraph, warn = TRUE)
}
\arguments{
\item{stratigraph}{A \code{stratigraph} object (see \code{\link[=stratigraph]{stratigraph()}}).}

\item{warn}{Display warnings for invalid graphs? Default: \code{TRUE}.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Is an object a valid stratigraphic graph?
}
\details{
Checks whether a stratigraph object is a valid stratigraphic graph. Currently
looks for:
\itemize{
\item Whether the graph contains cycles
}
}
