% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_prune.R
\name{strg_prune}
\alias{strg_prune}
\title{Remove redundant relations from a stratigraph}
\usage{
strg_prune(strg)
}
\arguments{
\item{strg}{A \code{\link[=stratigraph]{stratigraph()}} object to prune.}
}
\value{
A \link{stratigraph} with redundant relations removed.
}
\description{
strg_prune() removes redundant relations from a stratigraphic
graph by computing its
\href{https://en.wikipedia.org/wiki/Transitive_reduction}{transitive reduction}.
The result is a 'pruned' graph that follows Harris' "Law of Stratigraphical
Succession": that only the uppermost and undermost relations are significant
when placing a unit in a stratigraphic sequence.
}
\examples{

bushy_stratigraphy <- stratigraph(
  tibble::tibble(
    id = letters[1:5],
    above = list(NA, "a", "a", c("a", "b", "c"), c("a", "c", "d"))
  ),
  "id", "above"
)

strg_prune(bushy_stratigraphy)
}
