% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_oxcal}
\alias{write_oxcal}
\title{Write CQL to a file}
\usage{
write_oxcal(cql, file)
}
\arguments{
\item{cql}{A \code{cql} object. See \code{\link[=cql]{cql()}}.}

\item{file}{Path to a file.}
}
\value{
Returns \code{cql} invisibly.
}
\description{
Writes a CQL script (from \code{\link[=cql]{cql()}}) to an .oxcal file, for input to OxCal.
}
\examples{
\dontrun{
script <- cql(
  cql_r_date("ABC-001", 10100, 50),
  cql_r_date("ABC-002", 10200, 50)
)
write_oxcal(script, tempfile(fileext = ".oxcal"))
}
}
\seealso{
Other CQL functions:
\code{\link[=cql]{cql()}},
\code{\link[=cql_boundary]{cql_boundary()}},
\code{\link[=cql_n]{cql_n()}},
\code{\link[=cql_options]{cql_options()}},
\code{\link{cql_other}},
\code{\link[=cql_phase]{cql_phase()}},
\code{\link[=cql_r_date]{cql_r_date()}},
\code{\link[=cql_sequence]{cql_sequence()}}
}
\concept{CQL functions}
\concept{write functions}
