% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql.R
\name{write_oxcal}
\alias{write_oxcal}
\title{Write CQL to a file}
\usage{
write_oxcal(cql, file)
}
\arguments{
\item{cql}{A \code{cql} object. See \code{\link[=cql]{cql()}}.}

\item{file}{Path to a file.}
}
\value{
Returns \code{cql} invisibly.
}
\description{
Writes CQL code from \code{\link[=cql]{cql()}} to an .oxcal file, for input to OxCal.
}
\seealso{
Other CQL functions: 
\code{\link{cql_boundary}()},
\code{\link{cql_options}()},
\code{\link{cql_other}},
\code{\link{cql_phase}()},
\code{\link{cql_r_date}()},
\code{\link{cql_sequence}()},
\code{\link{cql}()}
}
\concept{CQL functions}
